+++ src/main/java/net/frozenblock/wilderwild/datagen/recipe/WWNaturalRecipeProvider.java
@@ -34,6 +34,8 @@ import org.jetbrains.annotations.Nullable;
 public final class WWNaturalRecipeProvider {
 
 	static void buildRecipes(RecipeOutput exporter) {
+		oneToOneConversionRecipe(exporter, Items.YELLOW_DYE, WWBlocks.WILDFLOWERS, "yellow_dye");
+
 		oneToOneConversionRecipe(exporter, Items.RED_DYE, WWBlocks.RED_HIBISCUS, "red_dye");
 		oneToOneConversionRecipe(exporter, Items.YELLOW_DYE, WWBlocks.YELLOW_HIBISCUS, "yellow_dye");
 		oneToOneConversionRecipe(exporter, Items.WHITE_DYE, WWBlocks.WHITE_HIBISCUS, "white_dye");